fix(logout); cannot logout properly due to wrongly placed middleware and refactor code
