+++ src/routes/index.ts
@@ -1,10 +1,8 @@
 import authMiddleware from '~/middlewares/auth-middleware';
 import { Hono } from "hono";
-import { deleteCookie, getCookie } from 'hono/cookie';
 
 import { getEvents } from '~/utils/event';
 import { EventProps } from '~/types/event';
-import { deleteToken } from '~/services/auth-service';
 
 import authRoute from "./auth";
 import eventRoute from "./event";
@@ -43,27 +41,6 @@ const app = new Hono()
         }),
       })
     );
-  })
-  .get('/logout', async (c) => {
-    const accessToken = getCookie(c, 'access_token');
-    const refreshToken = getCookie(c, 'refresh_token');
-
-    if (accessToken) {
-      deleteCookie(c, 'access_token');
-    }
-
-    if (refreshToken) {
-      try {
-        await deleteToken({ token: refreshToken });
-        deleteCookie(c, 'refresh_token');
-      } catch (error: unknown) {
-        deleteCookie(c, 'refresh_token');
-      }
-    }
-
-    // Using 'HX-Redirect' so that HTMX 'hx-get' can redirect into another page
-    c.header('HX-Redirect', '/login');
-    return c.body(null);
   })
   .route('/auth', authRoute)
   .route('/event', eventRoute)
